feat: implement new recap video generating api using aws lambda
