+++ photo-service/src/main/java/kr/mafoo/photo/controller/dto/request/RecapCreateRequestOld.java
@@ -1,0 +1,12 @@
+package kr.mafoo.photo.controller.dto.request;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+import kr.mafoo.photo.annotation.ULID;
+
+@Schema(description = "리캡 생성 요청")
+public record RecapCreateRequestOld(
+        @ULID
+        @Schema(description = "앨범 ID", example = "test_album_id")
+        String albumId
+) {
+}
