+++ photo-service/src/main/java/kr/mafoo/photo/controller/dto/response/RecapResponse.java
@@ -1,6 +1,7 @@
 package kr.mafoo.photo.controller.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
+import kr.mafoo.photo.service.dto.RecapUrlDto;
 
 @Schema(description = "리캡 응답")
 public record RecapResponse(
@@ -14,6 +15,14 @@ public record RecapResponse(
         ) {
                 return new RecapResponse(
                         recapUrl
+                );
+        }
+
+        public static RecapResponse fromDto(
+            RecapUrlDto dto
+        ) {
+                return new RecapResponse(
+                    dto.recapUrl()
                 );
         }
 }
