+++ photo-service/src/main/java/kr/mafoo/photo/service/dto/RecapUrlDto.java
@@ -1,0 +1,7 @@
+package kr.mafoo.photo.service.dto;
+
+public record RecapUrlDto(
+    String recapUrl
+) {
+
+}
